app/requirements.txt
pandas==1.3.*
st_annotated_text==1.0.1
streamlit==1.*
quote==2.*
protobuf==3.19.*
measurement==3.*
